refactor : 패키지 구조 변경
도메인 영역과 입출력 영역의 분리를 위해 패키지 구조를 변경했다.
