+++ src/main/java/menu/domain/menu/MenuGenerator.java
@@ -1,4 +1,4 @@
-package menu.menu;
+package menu.domain.menu;
 
 import java.util.HashMap;
 import java.util.List;
